requirements.txt
fastapi
uvicorn
httpx
python-dotenv
openai
elevenlabs
google-cloud-storage
